Rewrote dag loading.
