+++ gettsim/dag.py
@@ -75,13 +75,8 @@ def compute_taxes_and_transfers(
         "benefits/kinderzuschlag_dag.py",
         "benefits/unterhalt_dag.py",
         "benefits/wohngeld_dag.py",
-        "taxes/kindergeld.py",
         "renten_anspruch_dag.py",
-        "taxes/favorability_check.py",
-        "taxes/eink_st.py",
-        "taxes/zu_versteuerndes_eink_dag.py",
-        "taxes/soli_st.py",
-        "taxes/abgelt_st.py",
+        "taxes",
     ]
     for file in internal_function_files:
         new_funcs = load_functions(Path(__file__).parent / file)
